backend/requirements_temp.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
boto3==1.42.53
botocore==1.42.53
cbor2==5.8.0
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
distro==1.9.0
Django==5.2.11
django-cors-headers==4.9.0
django-environ==0.13.0
django-ratelimit==4.1.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.3.1
requests==2.32.5
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
yarl==1.22.0
zope.interface==8.2
